replacing double with BigDecimal for the value of price, creating an additional enum

+++ src/main/java/ru/aston/karpenko_ds/task1/service/Sales.java
@@ -1,11 +1,12 @@
 package ru.aston.karpenko_ds.task1.service;
 
+import java.math.BigDecimal;
 import java.util.Comparator;
 import java.util.List;
 import java.util.function.ToDoubleFunction;
 
 public interface Sales<T> {
-    double totalPrice(ToDoubleFunction<T> function);
+    BigDecimal totalPrice(ToDoubleFunction<T> function);
     List<T> sortSalesList(Comparator<T> comparator);
     Sales<T> add(T sale);
     T getSale(int index);
